Integrate with CartConf as a standalone Avocado varianter
This is a specialized varianter and a separate project providing
Cartesian configuration reimplemented in Rust. It is still at
purely experimental stage and thus entirely optional.

Use the CartConf Avocado plugin as dependency for Avocado VT where
identical functionality exists and reimplement locally where such
does not. This is optional and experimental dependency which will
fall back to the local cartesian_config module but users would
still be advised to try it out if they end up importing the local
Cartesian parser blob within the VT plugin.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>